+++ src/libertem/corrections/__init__.py
@@ -1,3 +1,3 @@
-from .corrset import CorrectionSet
+from libertem.io.corrections.corrset import CorrectionSet
 
 __all__ = ["CorrectionSet"]